change type of read to ssize_t

+++ src/socket/CGI.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   CGI.cpp                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nouchata <nouchata@student.42.fr>          +#+  +:+       +#+        */
+/*   By: tmatis <tmatis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/18 17:38:29 by nouchata          #+#    #+#             */
-/*   Updated: 2021/10/26 07:34:17 by nouchata         ###   ########.fr       */
+/*   Updated: 2021/10/26 19:16:30 by tmatis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -244,7 +244,7 @@ bool			CGI::get_response(int const &revents)
 	if (revents & POLLIN || this->_response_flag)
 	{
 		char		buffer[1024];
-		int			i;
+		ssize_t			i;
 		bool		parsecgi_ret = false;
 
 		memset(buffer, 0, 1024);
